feat(deploy): read upgrade command from config
We don't want every library to know the special commands for each repo
it deploys to.  Instead, it seems good that we would have a config
file that allows us to specify the deploy patterns for that repo.

Currently this just supports upgrade commands, but will allow us to
remove the awkward config file in the settings.