feat(deploy): prevent major versions from deploy
We want to take a more manual approach to major version upgrades.
This means that we want to check if we are attempting
to do a major upgrade
and if so, abort the change.

This was important to do at a per consumer level so that apps that were
behind for 1 reason or another would not be accidentally upgraded on
a patch release shortly after a major release.