feat: raise errors when versions are not supported
There were scenarios where an upgrade was not compatible because of
peerDependencies not being valid.  This caused the version to be
upgraded to a lower version than what we intended.  Instead
of allowing that to happen, we will now raise an error that communicates
why we were not able to upgrade to the newer version.