fix: do not auto-deploy major RCs
because major versions contain breaking changes, there is a high
probability that deploying RCs will break staging environments.
Because of that, let's remove the deploy ability for release candidates
for major updates.

This change does not apply to qa builds, which push to protonova
environments.  Those can still be useful for creating an environment to
explore the impacts of a major change on an isolated environment.